refactor: remove "security" prefix
