+++ usecases/blea-gov-base-ct/lib/construct/detection.ts
