+++ usecases/blea-gov-base-ct/lib/stack/blea-gov-base-ct-via-service-catalog-stack.ts
@@ -1,9 +1,9 @@
 import { Stack, StackProps } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import { CloudFormationProduct, Portfolio, ProductStack, ProductStackHistory } from 'aws-cdk-lib/aws-servicecatalog';
-import { Iam } from '../construct/security-iam';
-import { Logging } from '../construct/security-logging';
-import { Detection } from '../construct/security-detection';
+import { Iam } from '../construct/iam';
+import { Logging } from '../construct/logging';
+import { Detection } from '../construct/detection';
 import { Role } from 'aws-cdk-lib/aws-iam';
 
 export interface BLEAGovBaseCtScStackProps extends StackProps {
