+++ usecases/blea-gov-base-standalone/lib/construct/detection.ts
